% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dosing_schedule_df}
\alias{dosing_schedule_df}
\title{The dosing schedule data frame.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
dosing_schedule_df
}
\description{
A data frame with 4 rows and 4 columns:
\describe{
  \item{\code{drug}}{The numeric code of the drug.}
  \item{\code{target_days}}{The target number of days per treatment cycle.}
  \item{\code{target_kits}}{The target number of kits per treatment cycle.}
  \item{\code{max_cycles}}{The maximum number of treatment cycles.}
}
}
\keyword{datasets}
