% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_description_df}
\alias{drug_description_df}
\title{The drug description data frame.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
drug_description_df
}
\description{
A data frame with 4 rows and 3 columns:
\describe{
\item{\code{drug}}{The numeric code of the drug.}
\item{\code{drug_name}}{The name of the drug.}
\item{\code{dose_unit}}{The dose unit for drug dispensing.}
}
For drug demand forecasting, the default dose unit is "kit" for all drugs.
}
\keyword{datasets}
