% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dose_pp.R
\name{f_cum_dose}
\alias{f_cum_dose}
\title{Cumulative Dose}
\usage{
f_cum_dose(x, w, d, N)
}
\arguments{
\item{x}{Treatment duration.}

\item{w}{Number of days per treatment cycle.}

\item{d}{Dose per treatment cycle.}

\item{N}{Maximum number of treatment cycles.}
}
\value{
The cumulative dose to dispense for the drug over a specified
treatment duration.
}
\description{
Obtains the cumulative dose given treatment duration and
dosing schedule.
}
\examples{

f_cum_dose(c(28, 70), 21, 2, 10000)

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
