% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_drug_demand.R
\name{f_treatment_by_drug_df}
\alias{f_treatment_by_drug_df}
\title{Drug to Treatment Mapping}
\usage{
f_treatment_by_drug_df(treatment_by_drug, drug_name, dose_unit)
}
\arguments{
\item{treatment_by_drug}{The indicator matrix of treatment by drug
combinations.}

\item{drug_name}{The name of the drug.}

\item{dose_unit}{The dose unit used for drug dispensing.}
}
\value{
A data frame indicating the treatments
associated with each drug, including the following variables:
\code{treatment}, \code{drug}, \code{drug_name}, and
\code{dose_unit}.
}
\description{
Obtains a data frame that indicates the treatment(s)
associated with each drug.
}
\examples{

drug_name = drug_description_df$drug_name
dose_unit = drug_description_df$dose_unit
treatment_by_drug_df <- f_treatment_by_drug_df(
  treatment_by_drug, drug_name, dose_unit)
treatment_by_drug_df

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
