% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kit_description_df}
\alias{kit_description_df}
\title{The kit description data frame.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.
}
\usage{
kit_description_df
}
\description{
A data frame with the following columns:
\describe{
\item{\code{drug}}{The numeric code of the drug.}
\item{\code{drug_name}}{The name of the drug.}
\item{\code{kit}}{The numeric code of the kit type.}
\item{\code{kit_name}}{The name of the kit type.}
\item{\code{p_kit}}{The prior probability of different kit types
within a drug.}
}
}
\keyword{datasets}
