% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_drug_demand.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random Number Generator for the Dirichlet Distribution}
\usage{
rdirichlet(n = 1, alpha)
}
\arguments{
\item{n}{The number of observations.}

\item{alpha}{The shape parameters of the Dirichlet distribution.}
}
\value{
A matrix of n rows and k columns, where n is the number of
observations and k is the number of cells.
}
\description{
Generates cell probabilities from the Dirichlet distribution.
}
\examples{

rdirichlet(2, c(50, 20, 30))

}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
