% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_by_drug}
\alias{treatment_by_drug}
\title{The indicator matrix of treatment by drug combinations.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 4 columns.
}
\usage{
treatment_by_drug
}
\description{
A matrix with dimensions k x l, where k equals 3 representing the
number of treatment groups, and l equals 4 representing the number
of drugs. In this matrix, a value of 1 signifies the presence of
the drug within a treatment group, while a value of 0 indicates
the absence of the drug in that particular treatment group.
}
\keyword{datasets}
