% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_by_drug_df}
\alias{treatment_by_drug_df}
\title{The data frame indicating the treatments associated with each drug.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
treatment_by_drug_df
}
\description{
A data frame with the following columns:
\describe{
\item{\code{treatment}}{The numeric code of the treatment group.}
\item{\code{drug}}{The numeric code of the drug.}
}
}
\keyword{datasets}
