% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visitview1}
\alias{visitview1}
\title{The observed subject drug dispensing data before enrollment completion.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2290 rows and 8 columns.
}
\usage{
visitview1
}
\description{
A data frame with 2290 rows and 8 columns:
\describe{
  \item{\code{usubjid}}{The unique subject ID.}
  \item{\code{visit}}{The drug dispensing visit, e.g. "Cycle 1 Day 1".}
  \item{\code{date}}{The date of the drug dispensing visit.}
  \item{\code{drug}}{The numeric code of the drug.}
  \item{\code{drug_name}}{The name of the the drug.}
  \item{\code{dose_unit}}{The dose unit for drug dispensing.}
  \item{\code{kit_number}}{The kit number for drug dispensing.}
  \item{\code{dispensed_quantity}}{The number of kits dispensed
  at the visit.}
}
}
\keyword{datasets}
